Changed concurrency test to use Threading.
Unfortunately the easiest solution (futures module) is not part of
Python 2 so I had to change to threading for this test.

Also the queue module in Python 2 is named Queue so had to use
try/except to make sure it works for it.

The test puts a delay in obtaining the response which with the old
way of instantiating of the Profiler causes immediate exception.

The test checks if number of results from the thread queue matches
the number of repeated calls to the middleware.

We no longer need a very large number of calls made to trigger the error
as the sleep fixes it in place.